requirements_cloudwatch.txt
boto3>=1.26.0
pyyaml>=6.0
pandas>=1.5.0
matplotlib>=3.6.0
openpyxl>=3.0.0
